 \name{gskat-internal} 
        \alias{Meff_PCA} 
		\alias{PCA_cutoff}
		\alias{blockMatrixDiagonal}
		\alias{fit_FSKAT_IC}
		\alias{fit_FSKAT_IC_cont}
		\alias{frechet}
		\alias{gee_wrap}
		\alias{geeglm_wrap}
		\alias{genoC}
		\alias{gskat_fit}
		\alias{inferCutoff}
		\alias{kindepth}
		\alias{kinship}
		\alias{mybin}
		\alias{myfit.RFAM_IC}
		\alias{mykin}
		\alias{mysqrt}
		\alias{simpleM}
        \title{Internal gskat Functions} 
        \description{ 
          Internal mypackage functions 
        } 
        \details{ 
          These are not to be called by the user. 
        } 
\keyword{ internal } 