\name{gskat-package}
\alias{gskat-package}
\alias{gskat}
\docType{package}
\title{
Family based association test via kernel machine regression score test
}
\description{R package for GEE_SKAT.}
\details{
\tabular{ll}{
Package: \tab gskat\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-12-30\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}

\author{
XUEFENG WANG
Maintainer:  Xuefeng Wang <pwxfor@gmail.com>
}

\keyword{ package }